Fix more (MSVC) compiler warnings
reg. `size_t` vs. `int`, still not all (and some of these changes result in new warnings elsewhere), but 496 less than before (still 452 left).

+++ src/engraving/libmscore/score.cpp
@@ -4554,9 +4554,9 @@ EngravingItem* Score::getScoreElementOfMeasureBase(MeasureBase* mb) const
 //   nmeasure
 //---------------------------------------------------------
 
-int Score::nmeasures() const
+size_t Score::nmeasures() const
 {
-    int n = 0;
+    size_t n = 0;
     for (const Measure* m = firstMeasure(); m; m = m->nextMeasure()) {
         n++;
     }
